requirements.txt
streamlit==1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-text-splitters>=1.0.0
chromadb==0.4.22
pypdf2==3.0.1
python-dotenv==1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
datasets>=4.5.0
pyarrow>=21.0.0
